feat: adding routes authentication

+++ server/src/routes/memories.ts
@@ -1,11 +1,20 @@
 import { FastifyInstance } from "fastify";
 import { z } from 'zod'
 import { prisma } from "../lib/prisma";
+import { REPL_MODE_SLOPPY } from "repl";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
+
+
+    app.get('/memories', async (request) => {
+
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub,
+            },
             orderBy: {
                 createdAt: 'asc',
             }
@@ -20,7 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         })
     })
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -32,6 +41,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id,
             }
         })
+
+        if(!memory.isPublic && memory.userId !== request.user.sub) {
+            return reply.status(401).send('')
+        }
 
         return memory
     })
@@ -50,7 +63,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: '5fef4724-f3e2-4e7a-8d14-62b389e736d7'
+                userId: request.user.sub,
             }
         })
 
@@ -58,7 +71,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
     })
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const bodySchema = z.object({
             content: z.string(),
             coverUrl: z.string(),
@@ -72,7 +85,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
         const { id } = paramsSchema.parse(request.params)
         const {content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if(memory.userId !== request.user.sub) {
+            return reply.status(401).send('')
+        }
+
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -86,12 +109,23 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
     })
 
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
 
         const { id } = paramsSchema.parse(request.params)
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if(memory.userId !== request.user.sub) {
+            return reply.status(401).send('')
+        }
+
 
         await prisma.memory.delete({
             where: {
